+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-XX-XX
+### Added
+- Add new directives : "media-src" 
+- Add a field to input "report-to"
+
 ### Changed
 - Add patches & multivalue_form_element to composer.json